requirements.txt
uvicorn==0.12.2
fastapi==0.61.1
aiofiles==0.5.0
Pillow==8.0.0
qrcode==6.1
Jinja2==2.11.2
websockets==8.1
imutils==0.5.3
pandas==0.25.1


opencv-python
facenet_pytorch==2.4.1
